Implementação do metodo do joão na minha classe e ajeitada em algumas incompatibilidades

+++ src/lsystems/Desenho.java
@@ -11,10 +11,10 @@ package lsystems;
  */
 public class Desenho
 {
-    public static void desenhaSVG(int angulo, String end)
+    public static void desenhaSVG(double angulo, String end)
     {
         String Tratamento = end;
-        int angle = angulo;
+        double angle = angulo;
         StringBuilder sb = new StringBuilder();
     }
 }
